requirements.txt
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.22.0
azure-mgmt-apimanagement==5.0.0
azure-mgmt-core==1.5.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
